Move the ticket modal into a new component so you can click out of it
